Expenses Model,Repo,Controller,Expection

+++ src/main/java/com/example/simplecrm/repository/ExpensesClaimRepository.java
@@ -3,6 +3,6 @@ package com.example.simplecrm.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.example.simplecrm.model.ExpensesClaim;
 
-public interface ExpensesClaimRepository extends JpaRepository<ExpensesClaim, Integer> {
+public interface ExpensesClaimRepository extends JpaRepository<ExpensesClaim, Long> {
 
 }
